pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.0.0",
    "firebase-admin>=7.1.0",
    "google-cloud-firestore>=2.0.0",
    "grpcio>=1.60.0",
    "pytz>=2023.0",
]
